+++ src/main/java/com/pepero/jcb/api/exception/engine/UCIEngineException.java
@@ -7,4 +7,8 @@ public class UCIEngineException extends RuntimeException {
     public UCIEngineException(String message) {
         super(message);
     }
+
+    public UCIEngineException(String message, Throwable cause) {
+        super(message, cause);
+    }
 }